Keep connection partialy implemented

+++ login.js
@@ -3,6 +3,8 @@ const baseURL = getBaseUrl();
 $(document).ready(function(){
       let nav = montarNavBar();
       document.getElementById('navbanco').innerHTML = nav;
+      let flagKeepCon = document.getElementById('manterConectado');
+     
 
       document.getElementById('submit_login').addEventListener('click', () => {
             let usuario = document.getElementById('usuario').value,
@@ -16,9 +18,19 @@ $(document).ready(function(){
                   }).then( res => {
                         if (res.status == 200 ){
                               loader(false);
+                              /*
+                              Por falta de recursos da API, não é possível realmente implementar a funcionalidade
+                              de manter a conexão, pois o token possui uma validade limitada e não há como 
+                              reautenticar o usuário sem manter a senha salva, o que seria uma grave falha de segurança
+                              */
+                             if(flagKeepCon.checked){
+                                   localStorage.setItem('@token', res.data.token)
+                                   localStorage.setItem('userDataAccount', JSON.stringify(res.data))
+                              }else{
+                                    sessionStorage.setItem('@token', res.data.token)
+                                    sessionStorage.setItem('userDataAccount', JSON.stringify(res.data))
+                              }
                               window.location.replace('Dashboard.html')
-                              localStorage.setItem('@token', res.data.token)
-                              localStorage.setItem('userDataAccount', JSON.stringify(res.data))
                         }
